+++ src/launcher/java/org/truffleruby/launcher/RubyLauncher.java
@@ -339,7 +339,6 @@ public class RubyLauncher extends AbstractLanguageLauncher {
         out.println("TruffleRuby:");
         out.println("  -Xlog=SEVERE,WARNING,INFO,CONFIG,FINE,FINER,FINEST");
         out.println("                  set the TruffleRuby logging level");
-        out.println("  -Xoptions       print available TruffleRuby options");
         out.println("  -Xname=value    set a TruffleRuby option (omit value to set to true)");
         out.println("  -J-option=value Translates to --jvm.option=value");
     }